lesson-8-agentic-rag-with-python-and-cosmos-db/exercises/starter/requirements.txt
# Core dependencies
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.11.7

# Semantic Kernel
semantic-kernel==1.36.1

# Azure Cosmos DB dependencies
azure-cosmos==4.5.1
azure-identity==1.15.0

# External API integration dependencies
urllib3==2.0.7
httpx==0.25.2

# Data processing for API responses (Python 3.13 compatible versions)
numpy>=2.3.2
pandas>=2.3.1

# Optional: For better performance and caching
cachetools==5.3.2
